Move graph models to pkg/models
