+++ gqlgen.yml
@@ -35,7 +35,7 @@ struct_tag: gqlgen
 # gqlgen will search for any type names in the schema in these go packages
 # if they match it will use them, otherwise it will generate them.
 autobind:
-  - 'github.com/cmsgov/easi-app/pkg/graph/model'
+  - 'github.com/cmsgov/easi-app/pkg/models'
 
 # This section declares type mapping between the GraphQL and go type systems
 #
@@ -56,4 +56,4 @@ models:
       - github.com/99designs/gqlgen/graphql.Int32
   UUID:
     model:
-      - github.com/cmsgov/easi-app/pkg/graph/model.UUID
+      - github.com/cmsgov/easi-app/pkg/models.UUID
