+++ pkg/graph/schema.resolvers.go
@@ -11,10 +11,11 @@ import (
 
 	"github.com/cmsgov/easi-app/pkg/graph/generated"
 	"github.com/cmsgov/easi-app/pkg/graph/model"
+	"github.com/cmsgov/easi-app/pkg/models"
 )
 
 func (r *mutationResolver) CreateAccessibilityRequest(ctx context.Context, input *model.CreateAccessibilityRequestInput) (*model.CreateAccessibilityRequestPayload, error) {
-	request, err := r.store.CreateAccessibilityRequest(ctx, &model.AccessibilityRequest{
+	request, err := r.store.CreateAccessibilityRequest(ctx, &models.AccessibilityRequest{
 		Name: input.Name,
 	})
 	if err != nil {
@@ -27,7 +28,7 @@ func (r *mutationResolver) CreateAccessibilityRequest(ctx context.Context, input
 	}, nil
 }
 
-func (r *queryResolver) AccessibilityRequest(ctx context.Context, id uuid.UUID) (*model.AccessibilityRequest, error) {
+func (r *queryResolver) AccessibilityRequest(ctx context.Context, id uuid.UUID) (*models.AccessibilityRequest, error) {
 	return r.store.FetchAccessibilityRequestByID(ctx, id)
 }
