+++ pkg/models/uuid.go
@@ -1,4 +1,4 @@
-package model
+package models
 
 import (
 	"errors"
