+++ pkg/storage/accessibility_request.go
@@ -10,12 +10,11 @@ import (
 
 	"github.com/cmsgov/easi-app/pkg/appcontext"
 	"github.com/cmsgov/easi-app/pkg/apperrors"
-	"github.com/cmsgov/easi-app/pkg/graph/model"
 	"github.com/cmsgov/easi-app/pkg/models"
 )
 
 // CreateAccessibilityRequest adds a new accessibility request in the database
-func (s *Store) CreateAccessibilityRequest(ctx context.Context, request *model.AccessibilityRequest) (*model.AccessibilityRequest, error) {
+func (s *Store) CreateAccessibilityRequest(ctx context.Context, request *models.AccessibilityRequest) (*models.AccessibilityRequest, error) {
 	if request.ID == uuid.Nil {
 		request.ID = uuid.New()
 	}
@@ -51,8 +50,8 @@ func (s *Store) CreateAccessibilityRequest(ctx context.Context, request *model.A
 }
 
 // FetchAccessibilityRequestByID queries the DB for an accessibility matching the given ID
-func (s *Store) FetchAccessibilityRequestByID(ctx context.Context, id uuid.UUID) (*model.AccessibilityRequest, error) {
-	request := model.AccessibilityRequest{}
+func (s *Store) FetchAccessibilityRequestByID(ctx context.Context, id uuid.UUID) (*models.AccessibilityRequest, error) {
+	request := models.AccessibilityRequest{}
 
 	err := s.db.Get(&request, `SELECT * FROM accessibility_request WHERE id=$1`, id)
 	if err != nil {
@@ -72,8 +71,8 @@ func (s *Store) FetchAccessibilityRequestByID(ctx context.Context, id uuid.UUID)
 
 // FetchAccessibilityRequests queries the DB for an accessibility requests.
 // TODO implement cursor pagination
-func (s *Store) FetchAccessibilityRequests(ctx context.Context) ([]model.AccessibilityRequest, error) {
-	requests := []model.AccessibilityRequest{}
+func (s *Store) FetchAccessibilityRequests(ctx context.Context) ([]models.AccessibilityRequest, error) {
+	requests := []models.AccessibilityRequest{}
 
 	err := s.db.Select(&requests, `SELECT * FROM accessibility_request`)
 	if err != nil {
